20200605 修改部分

+++ ruoyi/src/main/java/com/ruoyi/project/party/domain/DjPartyTrainMember.java
@@ -38,7 +38,7 @@ public class DjPartyTrainMember extends BaseEntity
 
     /** 积分 */
     @Excel(name = "积分")
-    private Integer score;
+    private Double score;
 
     private Integer rankNum;
